+++ test/runtests.jl
@@ -33,8 +33,10 @@ end
         end
         include_test("groups/test_general_linear_group.jl")
         include_test("groups/test_orthogonal_group.jl")
+        include_test("groups/test_unitary_group.jl")
         include_test("groups/test_special_orthogonal_group.jl")
         include_test("groups/test_special_euclidean_group.jl")
+        include_test("groups/test_special_unitary_group.jl")
         include_test("groups/test_translation_group.jl")
     end
     include("test_aqua.jl")