+++ NEWS.md
@@ -16,7 +16,9 @@ Everything denoted by “formerly” refers to the previous name in [`Manifolds.
   * `TranslationGroup`
   * `SpecialEuclideanGroup` (formerly `SpecialEuclidean`)
   * `SpecialOrthogonalGroup` (formerly `SpecialOrthogonal`)
+  * `SpecialUnitaryGroup` (formerly `SpecialUnitary`)
   * `OrthogonalGroup` (formerly `Orthogonal`)
+  * `UnitaryGroup` (formerly `Unitary`)
   * `GeneralLinearGroup` (formerly `GeneralLinear`)
   * `LeftSemidirectProductLieGroup` (formerly `SemidirectProductGroup`)
   * `⋉` (alias for `LeftSemidirectProductGroupOperation` when a `default_left_action(G,H)` is defined for the two groups)